# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: AVANA API
  version: 1.0.0
  description: AVANA API contract (Sprint 1 foundations)
servers:
  - url: https://api.example.invalid
paths:
...
  /v1/organizations/{organizationId}/courses/{courseId}/documents/{documentId}/generate:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/DocumentId"
    post:
      operationId: generateContent
      summary: Synchronously generate AI content drafts for a document (mock provider, PR6-4)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentRequest"
      responses:
        "200":
          description: Generated content drafts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateContentResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Document not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Document is not in a generatable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/documents/{documentId}/generated:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/DocumentId"
    get:
      operationId: listGeneratedContent
      summary: List generated content drafts + citations for a document
      responses:
        "200":
          description: Generated content list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContentListResponse"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Document not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/documents/{documentId}/generated/{contentId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/GeneratedContentId"
    get:
      operationId: getGeneratedContent
      summary: Get a single generated content with citations
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContentResponse"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes: {}
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GeneratedContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: ["error", "request_id"]
      properties:
        request_id:
          type: string
        error:
          type: object
          additionalProperties: false
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: integer
                  - type: null
...
    DocumentStatus:
      type: string
      enum:
        - uploaded
        - pending_validation
        - validating
        - pending_extraction
        - extracting
        - pending_chunking
        - chunking
        - extracted
        - pending_generation
        - generating
        - review_pending
        - ready
        - failed
...
    GeneratedContentType:
      type: string
      description: Extensible union of AI-generated content types. Only "lesson" is enabled in PR6-4; flashcard/quiz/recommendation are added in later PRs.
      enum:
        - lesson
        - flashcard
        - quiz
        - recommendation
    GeneratedContentStatus:
      type: string
      description: AI artifact lifecycle (separate from the document processing lifecycle).
      enum:
        - draft
        - accepted
        - rejected
        - edited
        - regenerating
    CitationResource:
      type: object
      additionalProperties: false
      required: ["document_chunk_id"]
      properties:
        document_chunk_id:
          type: string
          format: uuid
    TokenUsage:
      type: object
      additionalProperties: false
      required: ["input_tokens", "output_tokens"]
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
    GeneratedContentResource:
      type: object
      additionalProperties: false
      required:
        - id
        - document_id
        - course_id
        - type
        - status
        - payload
        - prompt_version
        - model
        - token_usage
        - citations
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/GeneratedContentType"
        status:
          $ref: "#/components/schemas/GeneratedContentStatus"
        payload:
          type: object
          additionalProperties: true
          description: Typed content payload (discriminated by type).
        prompt_version:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        token_usage:
          $ref: "#/components/schemas/TokenUsage"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/CitationResource"
        created_at:
          type: string
          description: UTC ISO datetime
        updated_at:
          type: string
          description: UTC ISO datetime
    GenerateContentRequest:
      type: object
      additionalProperties: false
      properties:
        types:
          type: array
          description: Content types to generate. Defaults to ["lesson"] when omitted.
          items:
            $ref: "#/components/schemas/GeneratedContentType"
        prompt_version:
          type: string
          description: Prompt version used for regeneration/cost analysis.
    GenerateContentResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "contents", "document_status"]
      properties:
        request_id:
          type: string
        contents:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedContentResource"
        document_status:
          $ref: "#/components/schemas/DocumentStatus"
    GeneratedContentListResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "contents"]
      properties:
        request_id:
          type: string
        contents:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedContentResource"
    GeneratedContentResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "content"]
      properties:
        request_id:
          type: string
        content:
          $ref: "#/components/schemas/GeneratedContentResource"

    # -------------------------------------------------------------------
    # PR6-6 Human review & acceptance contract types
    # -------------------------------------------------------------------
...